moves the demo pages to the Pages folder
